Leverage <details open> attribute for handling state
