+++ lib/adoptoposs_web/templates/project/sharing.html.eex
@@ -1,5 +1,5 @@
 <details class="toggle">
-  <summary class="toggle-trigger icon-button subtle my-2" title="Share" phx-hook="SharingToggle">
+  <summary class="toggle-trigger icon-button subtle my-2" title="Share" phx-hook="SharingReset">
     <svg xmlns:svg="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
       <path d="m 16,13 v 4 c 0,0 -0.9,2 -2,2 H 6 C 4.9,19 4,18.1 4,17 v -4 h 2 v 4 h 8 V 13 Z M 9,4.83 v 9.492034 h 2 V 4.83 L 14.07,7.9 15.49,6.49 10,1 9.3,1.7 4.5,6.5 5.92,7.9 9,4.84 Z"/>
     </svg>
